Implemented /tiles.json
